Change to floats

+++ src/constants/indicators.js
@@ -638,12 +638,12 @@ export const INDICATOR_SCHEME_ORDER = [
 ];
 
 export const INDICATOR_THRESHOLD_TRANSFORMERS = {
-  bws_cat: threshold => parseInt(threshold) ? parseInt(threshold) / 100 : 0,
-  bwd_cat: threshold => parseInt(threshold) ? parseInt(threshold) / 100 : 0,
-  gtd_cat: threshold => parseInt(threshold) || 0,
-  cep_cat: threshold => parseInt(threshold) || 0,
-  udw_cat: threshold => parseInt(threshold) ? parseInt(threshold) / 100 : 0,
-  usa_cat: threshold => parseInt(threshold) ? parseInt(threshold) / 100 : 0
+  bws_cat: threshold => parseFloat(threshold) ? parseFloat(threshold) / 100 : 0,
+  bwd_cat: threshold => parseFloat(threshold) ? parseFloat(threshold) / 100 : 0,
+  gtd_cat: threshold => parseFloat(threshold) || 0,
+  cep_cat: threshold => parseFloat(threshold) || 0,
+  udw_cat: threshold => parseFloat(threshold) ? parseFloat(threshold) / 100 : 0,
+  usa_cat: threshold => parseFloat(threshold) ? parseFloat(threshold) / 100 : 0
 };
 
 export const EXCLUSIVE_MONTHLY_INDICATORS = ['bws_cat', 'bwd_cat', 'iav_cat'];
